server/requirements.txt
torch --extra-index-url https://download.pytorch.org/whl/cpu
torchvision --extra-index-url https://download.pytorch.org/whl/cpu
torchaudio --extra-index-url https://download.pytorch.org/whl/cpu
transformers
langchain
langchain-huggingface
langchain-google-genai
langchain-community
fastapi
pydantic
uvicorn
motor
python-jose[cryptography]
easyocr
passlib
matplotlib
numpy
opencv_contrib_python
opencv_python_headless
pymongo
python-dotenv
python-multipart
pydantic-settings
bcrypt